+++ lib/src/programs/memo/program.dart
@@ -2,8 +2,8 @@
 /// ------------------------------------------------------------------------------------------------
 
 import 'dart:convert';
-import 'package:solana_common/utils/library.dart';
 import 'package:solana_common/utils/types.dart';
+import 'package:solana_common/utils/utils.dart';
 import '../../../programs/program.dart';
 import '../../../src/public_key.dart';
 import '../../../src/transaction/transaction.dart';