+++ test/utils.dart
@@ -9,7 +9,7 @@ import 'package:solana_common/utils/buffer.dart';
 /// Test Utilities
 /// ------------------------------------------------------------------------------------------------
 
-/// Max/Min values
+// Max/Min values
 const int minUInt = 0;
 const int maxUint8 = 255;
 const int maxUint16 = 65535;
@@ -21,6 +21,7 @@ final Buffer minInt64Buffer = Buffer.fromList([0, 0, 0, 0, 0, 0, 0, 128]);
 const int maxInt64 = 9223372036854775807;
 final Buffer maxInt64Buffer = Buffer.fromList([255, 255, 255, 255, 255, 255, 255, 127]);
 
+// A test wallet account.
 final web3.Keypair wallet = web3.Keypair.fromSeedSync(
   Uint8List.fromList([
     183, 69, 110, 126, 141, 167, 148, 129, 79, 147, 153, 132, 80, 178, 86, 134, 
@@ -28,7 +29,7 @@ final web3.Keypair wallet = web3.Keypair.fromSeedSync(
   ])
 );
 
-/// Helpers
+// Helpers
 assertRange(final int value, { final int min = 0, final int max = 255 }) {
   assert(value >= min && value <= max);
 }